ENGR00136869: SATA device can't init and a typo in sata help
SATA device can't init and a typo in sata help.
Add delay in sata detect proceduce.

Currently, I have met 3 problems for this issue.
1. Seagate HD. It needs 1000 for timeout.
2. Hitachi HD. It needs 10000 for timeout.
3. In sata env case, it needs 100000 for timeout.

10000000 for timeout is just to avoid a dead loop,
And suppose this timeout should be enough for all normal
case.
It doesn't mean all HD need to wait this long time,
If tfd is ok, the loop will be breaked immediately.

Signed-off-by: Terry Lv <r65388@freescale.com>